+++ NEWS.md
@@ -7,6 +7,7 @@
 - shinyplot generic function.
 - Update parameters fHDmin and fHDmax.
 - New vignette 'IFNEvaluation'
+- New parameter 'WUE_decay' for reduction of relative WUE in Granier's model
 
 # Version 2.4.0
 - Functions 'spwb_stress', 'spwb_resistances' and 'spwb_waterUseEfficiency' renamed to 'droughtStress', 'resistances' and 'waterUseEfficiency', respectively, since they can now be applied to the output of several simulation function.